🏗️ Create ProfileItem mapper class
